Four waves: labels, rivers, urban 17a -- and five blockers that were never real
Twenty agent lanes across four verified waves. cargo test --workspace goes
2734 -> 2798 passed, 0 failed, 21 ignored.

Built
- CARTO labels, all three ruled steps: a label_class field on MapLabel, a
  generated labelling pass over five sources, and a per-class typography record
  transcribed from parts.js:363. The Water class had no entity that could ever
  be labelled, so lake_features() and a FeatureKind::Lake namer were added on
  their own RNG stream.
- The river entity: get_rivers() plus viewport hit-testing, which finally makes
  right_dock.gd's River context reachable.
- Landmark M8's five way-graph kinds, taking kinds() from 15 buildable to 20 of
  50. Each verified placing on a real generate_terrain world, not by flipping a
  flag. JUNCTION_MIN_WAYS corrected 3 -> 2, its inherited rationale shown false.
- Urban 17a golden-verified, closing the last non-16 urban milestone.
- Five re-entrant civ stages and a civ PARAMS group; a Units formatter with
  nautical miles; TERRAIN_APPEARANCE_SCOPE 16 and 19.
- The v2.11 re-freeze. v2.10 is deliberately kept and byte-unmodified, so every
  scope-document line range still resolves; the drift report is what a later
  pass uses to move them.

Parity defects found
- slope_at used f64::hypot where the reference uses Math.hypot, plus two more in
  the hydrology channel code. A third survives at enforce_channel_descent and is
  recorded rather than quietly fixed -- it feeds the terrain field directly.
- um_site_profile clamped the resource-context centre; mag carried a lower clamp
  the reference does not have.
- A fixture bug caught before it could be committed as truth.

Blockers that were never real
The "no JS runtime in this environment" claim gated three test headers and cost
a dispatched wave. node has been available for weeks, and everything the claim
gated shipped on 2026-08-15. tools/jsruntime_probe.js now proves both halves:
that the runtime executes the frozen reference, and that the committed fixtures
are its real output rather than the port's.

Four more document defects fixed the same way: a Cargo.toml authorisation still
advertised as live after the owner withdrew it, a ruling paraphrased into its
opposite, a landmark denominator that was never right, and a count section
reproducing the exact self-disagreement it exists to record. OUTSTANDING_WORK
goes 136 -> 128, every figure now derived by one script so the table and the
prose beside it cannot drift apart.

+++ cartalith-native/crates/cartalith-civ/src/trade.rs
@@ -1241,7 +1241,7 @@ pub fn civ_salt_access(w: &PlaceWorld, x: usize, y: usize, nav: NavKind) -> Salt
     if w.field.len() == n && w.res.salt.len() == n {
         let radius = js_max(3.0, js_round(w.gw as f64 / 128.0)) as usize;
         let mean =
-            civ_place_resource_context(w.res, w.field, w.gw, w.gh, w.sea, x, y, radius, false);
+            civ_place_resource_context(w.res, w.field, w.gw, w.gh, w.sea, x as i64, y as i64, radius, false);
         if mean.get("salt").copied().unwrap_or(0.0) > SALT_DEPOSIT_MEAN {
             return SaltAccess { has: true, source: "salt deposit" };
         }
@@ -1739,7 +1739,7 @@ pub fn civ_place_trade(
     let radius = js_max(3.0, js_round(world.gw as f64 / 128.0)) as usize;
     let mean: std::collections::HashMap<&str, f64> = if world.field.len() == n {
         civ_place_resource_context(
-            world.res, world.field, world.gw, world.gh, world.sea, x, y, radius, false,
+            world.res, world.field, world.gw, world.gh, world.sea, x as i64, y as i64, radius, false,
         )
     } else {
         std::collections::HashMap::new()
